Splitting NativeBufferPool into partial classes
